requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
aiofiles==24.1.0
python-multipart==0.0.9
pillow==10.4.0
